Added broadcasts

+++ processQueueTasks/closeTicket.js
@@ -10,36 +10,38 @@ module.exports = {
 		});
 
 		if (ticket && ticket.statusId === 100) {
-			// TODO: Change to broadcast
-			const channel = await client.channels.fetch(processQueueEntry.additions).catch(async () => {
-				//Nothing
-			});
+			client.shard.broadcastEval(async (c, { channelId }) => {
+				const channel = await c.channels.fetch(channelId).catch(async () => {
+					//Nothing
+				});
 
-			await channel.delete();
+				if (channel) {
+					await channel.delete();
 
-			// TODO: Change to broadcast
-			const guild = await client.guilds.fetch(channel.guild.id);
+					const guild = await client.guilds.fetch(channel.guild.id);
 
-			let openCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Open');
-			if (openCategory && !openCategory.children.first()) {
-				openCategory.delete();
-			}
-			let respondedCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Responded');
-			if (respondedCategory && !respondedCategory.children.first()) {
-				respondedCategory.delete();
-			}
-			let inActionCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - In Action');
-			if (inActionCategory && !inActionCategory.children.first()) {
-				inActionCategory.delete();
-			}
-			let awaitingResponseCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Awaiting Response');
-			if (awaitingResponseCategory && !awaitingResponseCategory.children.first()) {
-				awaitingResponseCategory.delete();
-			}
-			let closedCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Closed');
-			if (closedCategory && !closedCategory.children.first()) {
-				closedCategory.delete();
-			}
+					let openCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Open');
+					if (openCategory && !openCategory.children.first()) {
+						openCategory.delete();
+					}
+					let respondedCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Responded');
+					if (respondedCategory && !respondedCategory.children.first()) {
+						respondedCategory.delete();
+					}
+					let inActionCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - In Action');
+					if (inActionCategory && !inActionCategory.children.first()) {
+						inActionCategory.delete();
+					}
+					let awaitingResponseCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Awaiting Response');
+					if (awaitingResponseCategory && !awaitingResponseCategory.children.first()) {
+						awaitingResponseCategory.delete();
+					}
+					let closedCategory = guild.channels.cache.find(c => c.type === 'category' && c.name === 'Tickets - Closed');
+					if (closedCategory && !closedCategory.children.first()) {
+						closedCategory.delete();
+					}
+				}
+			}, { context: { channelId: processQueueEntry.additions } });
 		}
 
 		processQueueEntry.destroy();
